One step closer to calling the API
Bumped into issue #2.
